+++ NEWS.md
@@ -1,4 +1,4 @@
-# nanonext 0.10.0.9008 (development)
+# nanonext 0.10.0.9009 (development)
 
 #### New Features